% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_label}
\alias{add_label}
\title{Add value labels}
\usage{
add_label(variable, values)
}
\arguments{
\item{variable}{Variable to add labels to}

\item{values}{Named vector containing value-label pairs}
}
\value{
A labelled vector
}
\description{
Adds value labels to a variable
}
