% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_missings}
\alias{add_missings}
\title{Add value labels}
\usage{
add_missings(variable, missings)
}
\arguments{
\item{variable}{Variable to add missings to}

\item{missings}{Vector containg missing values}
}
\value{
A vector with user-defined missings
}
\description{
Adds missings to variable
}
