% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_vectors.R
\name{make_label_vector}
\alias{make_label_vector}
\title{Make label vector}
\usage{
make_label_vector(variable, values)
}
\arguments{
\item{variable}{Variable to create vector for}

\item{values}{Tibble containing value and label columns}
}
\value{
A named vector
}
\description{
Creates a named vector of value labels for all variables
}
