% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missings_vector.R
\name{make_missing_vector}
\alias{make_missing_vector}
\title{Make missings vector}
\usage{
make_missing_vector(variable, missings)
}
\arguments{
\item{variable}{Variable to create vector for}

\item{missings}{Tibble containg missing value information}
}
\value{
A vector of missing values
}
\description{
Creates a vector missing values for all variables
}
