% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_data.R
\name{request_data}
\alias{request_data}
\title{Request data from SoSciSurvey}
\usage{
request_data(URL, params)
}
\arguments{
\item{URL}{SoSciSurvey API URL}

\item{params}{Additional request parameters}
}
\value{
a list
}
\description{
Performs API request to SoSciSurvey
}
\keyword{internal}
