% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sosci.R
\name{sosci}
\alias{sosci}
\title{Load data from SoSciSurvey}
\usage{
sosci(URL, ...)
}
\arguments{
\item{URL}{SoSciSurvey API URL}

\item{...}{Additional parameters for API retrieval. See:
https://www.soscisurvey.de/help/doku.php/en:results:data-api}
}
\value{
A labelled tibble
}
\description{
Loads data from SoSciSurvey via the API and creates a labelled tibble.
}
